@@ -44,135 +44,132 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "6.1.0",
-    "@angular/cdk": "6.1.0",
-    "@angular/cli": "6.1.0",
-    "@angular/common": "6.1.0",
-    "@angular/compiler": "6.1.0",
-    "@angular/core": "6.1.0",
-    "@angular/forms": "6.1.0",
-    "@angular/http": "6.1.0",
-    "@angular/material": "6.1.0",
-    "@angular/platform-browser": "6.1.0",
-    "@angular/platform-browser-dynamic": "6.1.0",
-    "@angular/platform-server": "6.1.0",
-    "@angular/router": "6.1.0",
+    "@angular/animations": "^6.1.0",
+    "@angular/cdk": "^6.1.0",
+    "@angular/cli": "^6.1.0",
+    "@angular/common": "^6.1.0",
+    "@angular/compiler": "^6.1.0",
+    "@angular/core": "^6.1.0",
+    "@angular/forms": "^6.1.0",
+    "@angular/http": "^6.1.0",
+    "@angular/material": "^6.1.0",
+    "@angular/platform-browser": "^6.1.0",
+    "@angular/platform-browser-dynamic": "^6.1.0",
+    "@angular/platform-server": "^6.1.0",
+    "@angular/router": "^6.1.0",
     "@hmcts/frontend": "0.0.14-alpha",
-    "@hmcts/info-provider": "1.0.0",
-    "@hmcts/nodejs-healthcheck": "1.4.6",
-    "@hmcts/nodejs-logging": "3.0.0",
-    "@nguniversal/express-engine": "6.0.0",
-    "@nguniversal/module-map-ngfactory-loader": "6.0.0",
-    "applicationinsights": "1.0.3",
-    "@ng-toolkit/universal": "^1.1.45",
-    "base-64": "^0.1.0",
-    "body-parser": "1.18.3",
-    "cookie-parser": "1.4.3",
-    "core-js": "2.4.1",
-    "eslint-html-reporter": "0.5.2",
-    "exception-formatter": "1.0.7",
-    "express": "4.16.2",
-    "express-session": "1.15.6",
-    "govuk-frontend": "1.3.0",
-    "http-proxy-middleware": "0.18.0",
-    "https-proxy-agent": "^2.2.1",
-    "jsonpath": "1.0.0",
-    "jwt-decode": "2.2.0",
-    "lcov-result-merger": "3.1.0",
-    "lodash": "4.17.5",
-    "lodash.mapvalues": "4.6.0",
-    "log4js": "3.0.6",
-    "moment": "2.22.2",
-    "ngx-cookie": "4.0.2",
-    "node-fetch": "^2.2.0",
-    "otp": "0.1.3",
+    "@hmcts/info-provider": "^1.0.0",
+    "@hmcts/nodejs-healthcheck": "^1.4.6",
+    "@hmcts/nodejs-logging": "^3.0.0",
+    "@ng-toolkit/universal": "^1.1.47",
+    "@nguniversal/express-engine": "^6.0.0",
+    "@nguniversal/module-map-ngfactory-loader": "^6.0.0",
+    "applicationinsights": "^1.0.3",
+    "body-parser": "^1.18.3",
+    "cookie-parser": "^1.4.3",
+    "core-js": "^2.4.1",
+    "eslint-html-reporter": "^0.5.2",
+    "exception-formatter": "^1.0.7",
+    "express": "^4.16.2",
+    "express-session": "^1.15.6",
+    "govuk-frontend": "^1.3.0",
+    "http-proxy-middleware": "^0.18.0",
+    "jsonpath": "^1.0.0",
+    "jwt-decode": "^2.2.0",
+    "lcov-result-merger": "^3.1.0",
+    "log4js": "^3.0.6",
+    "moment": "^2.22.2",
+    "ngx-cookie": "^4.0.2",
+    "otp": "^0.1.3",
     "pdfjs-dist": "1.9.661",
-    "request": "2.87.0",
-    "request-promise": "4.2.2",
-    "request-promise-native": "1.0.5",
-    "rxjs": "6.2.2",
-    "rxjs-compat": "6.1.0",
-    "session-file-store": "1.2.0",
-    "stable": "0.1.8",
-    "ts-loader": "4.0.1",
-    "udk": "0.3.12",
-    "url": "0.11.0",
-    "util": "0.11.0",
-    "yarn": "1.6.0",
-    "zone.js": "0.8.19"
+    "request": "^2.85.0",
+    "request-promise": "^4.2.2",
+    "require": "^2.4.20",
+    "rxjs": "^6.2.2",
+    "rxjs-compat": "^6.1.0",
+    "session-file-store": "^1.2.0",
+    "stable": "^0.1.8",
+    "ts-loader": "^4.0.1",
+    "udk": "^0.3.12",
+    "url": "^0.11.0",
+    "util": "^0.11.0",
+    "yarn": "^1.6.0",
+    "zone.js": "^0.8.19"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "0.6.0",
-    "@angular-devkit/core": "0.6.0",
-    "@angular/compiler-cli": "6.1.0",
-    "@angular/language-service": "6.1.0",
-    "@hmcts/eslint-config": "1.4.0",
-    "@types/jasmine": "2.8.3",
-    "@types/jasminewd2": "2.0.2",
-    "@types/node": "6.0.60",
-    "argv": "0.0.2",
-    "autoprefixer": "7.2.3",
-    "chai": "4.1.2",
-    "chai-as-promised": "7.1.1",
-    "circular-dependency-plugin": "4.2.1",
-    "codacy-coverage": "^2.1.1",
-    "codecov": "^3.1.0",
-    "codelyzer": "4.0.1",
-    "concurrently": "3.6.1",
-    "copy-webpack-plugin": "4.4.1",
-    "cucumber": "4.2.1",
+    "@angular-devkit/build-angular": "^0.6.0",
+    "@angular-devkit/core": "^0.6.0",
+    "@angular/compiler-cli": "^6.1.0",
+    "@angular/language-service": "^6.1.0",
+    "@hmcts/eslint-config": "^1.4.0",
+    "@types/express": "^4.16.0",
+    "@types/jasmine": "~2.8.3",
+    "@types/jasminewd2": "~2.0.2",
+    "@types/node": "~6.0.60",
+    "@types/proxyquire": "^1.3.28",
+    "argv": "^0.0.2",
+    "autoprefixer": "^7.2.3",
+    "chai": "^4.1.2",
+    "chai-as-promised": "^7.1.1",
+    "circular-dependency-plugin": "^4.2.1",
+    "codacy-coverage": "^2.0.3",
+    "codecov": "^3.0.0",
+    "codelyzer": "^4.0.1",
+    "concurrently": "^3.6.1",
+    "copy-webpack-plugin": "~4.4.1",
+    "cucumber": "^4.2.1",
     "cucumber-html-report": "0.6.5",
-    "cucumber-html-reporter": "^4.0.4",
-    "cucumber-pretty": "1.4.4",
-    "cucumberjs-allure-reporter": "1.0.3",
-    "eslint": "4.19.1",
-    "eslint-config-standard": "12.0.0",
-    "eslint-plugin-import": "2.14.0",
-    "eslint-plugin-node": "7.0.1",
-    "eslint-plugin-promise": "4.0.1",
-    "eslint-plugin-standard": "4.0.0",
-    "file-loader": "1.1.5",
-    "form-data": "2.3.2",
-    "html-webpack-plugin": "2.29.0",
-    "istanbul-instrumenter-loader": "3.0.0",
-    "jasmine": "3.1.0",
-    "jasmine-core": "2.8.0",
-    "jasmine-reporters": "2.3.1",
-    "jasmine-spec-reporter": "4.2.1",
-    "json-server": "0.14.0",
-    "karma": "2.0.4",
-    "karma-chrome-launcher": "2.2.0",
-    "karma-coverage-istanbul-reporter": "1.2.1",
-    "karma-jasmine": "1.1.2",
-    "karma-jasmine-html-reporter": "0.2.2",
-    "karma-phantomjs-launcher": "1.0.4",
-    "less-loader": "4.0.5",
-    "minimist": "1.2.0",
-    "mocha": "5.2.0",
-    "mochawesome": "3.0.3",
-    "nodemon": "1.17.5",
-    "nsp": "3.2.1",
-    "nyc": "13.0.1",
-    "postcss-import": "11.0.0",
-    "postcss-loader": "2.0.10",
-    "postcss-url": "7.1.2",
-    "protractor": "5.4.1",
-    "protractor-cucumber-framework": "5.0.0",
+    "cucumber-html-reporter": "^4.0.2",
+    "cucumber-pretty": "^1.4.4",
+    "cucumberjs-allure-reporter": "^1.0.3",
+    "eslint": "^4.19.1",
+    "eslint-config-standard": "^12.0.0",
+    "eslint-plugin-import": "^2.14.0",
+    "eslint-plugin-node": "^7.0.1",
+    "eslint-plugin-promise": "^4.0.1",
+    "eslint-plugin-standard": "^4.0.0",
+    "file-loader": "^1.1.5",
+    "form-data": "^2.3.2",
+    "html-webpack-plugin": "^2.29.0",
+    "istanbul-instrumenter-loader": "^3.0.0",
+    "jasmine": "^3.1.0",
+    "jasmine-core": "~2.8.0",
+    "jasmine-reporters": "^2.3.1",
+    "jasmine-spec-reporter": "^4.2.1",
+    "json-server": "^0.14.0",
+    "karma": "^2.0.4",
+    "karma-chrome-launcher": "^2.2.0",
+    "karma-coverage-istanbul-reporter": "^1.2.1",
+    "karma-jasmine": "^1.1.2",
+    "karma-jasmine-html-reporter": "^0.2.2",
+    "karma-phantomjs-launcher": "^1.0.4",
+    "less-loader": "^4.0.5",
+    "minimist": "^1.2.0",
+    "mocha": "^5.2.0",
+    "mochawesome": "^3.0.3",
+    "nodemon": "^1.17.5",
+    "nsp": "^3.2.1",
+    "nyc": "^13.0.1",
+    "postcss-import": "^11.0.0",
+    "postcss-loader": "^2.0.10",
+    "postcss-url": "^7.1.2",
+    "protractor": "^5.4.1",
+    "protractor-cucumber-framework": "^5.0.0",
     "protractor-multiple-cucumber-html-reporter-plugin": "1.7.1",
-    "proxyquire": "2.0.1",
-    "raw-loader": "0.5.1",
-    "sass-loader": "6.0.6",
-    "should": "13.2.3",
-    "socks-proxy-agent": "3.0.1",
-    "sonar-scanner": "3.0.3",
-    "style-loader": "0.19.1",
-    "stylus-loader": "3.0.1",
-    "supertest": "3.1.0",
-    "ts-node": "4.1.0",
-    "tslint": "5.10.0",
-    "typescript": "2.7.2",
-    "uglifyjs-webpack-plugin": "1.1.8",
-    "url-loader": "0.6.2"
+    "proxyquire": "^2.0.1",
+    "raw-loader": "^0.5.1",
+    "sass-loader": "^6.0.6",
+    "should": "^13.2.3",
+    "socks-proxy-agent": "^3.0.1",
+    "sonar-scanner": "^3.0.3",
+    "style-loader": "^0.19.1",
+    "stylus-loader": "^3.0.1",
+    "supertest": "^3.1.0",
+    "ts-node": "~4.1.0",
+    "tslint": "^5.10.0",
+    "typescript": "2.9.2",
+    "uglifyjs-webpack-plugin": "^1.1.8",
+    "url-loader": "^0.6.2"
   },
   "snyk": true,
   "nyc": {
