@@ -1,19 +1,20 @@
-import {Inject, Injectable} from '@angular/core';
-import {CookieService} from 'ngx-cookie';
-import * as jwtDecode from 'jwt-decode';
-import {ConfigService} from '../config.service';
-import { RedirectionService} from '../routing/redirection.service';
+import { Inject, Injectable } from "@angular/core";
+import { CookieService } from "ngx-cookie";
+import * as jwtDecode from "jwt-decode";
+import { ConfigService } from "../config.service";
+import { RedirectionService } from "../routing/redirection.service";
 
 @Injectable({
-    providedIn: 'root'
+    providedIn: "root"
 })
 export class AuthService {
-
     COOKIE_KEYS;
 
-    constructor(private configService: ConfigService,
-                private cookieService: CookieService,
-                private redirectionService: RedirectionService) {
+    constructor(
+        private configService: ConfigService,
+        private cookieService: CookieService,
+        private redirectionService: RedirectionService
+    ) {
         this.COOKIE_KEYS = {
             TOKEN: this.configService.config.cookies.token,
             USER: this.configService.config.cookies.userId
@@ -23,7 +24,9 @@ export class AuthService {
     generateLoginUrl() {
         const base = this.configService.config.services.idam_web;
         const clientId = this.configService.config.idam_client;
-        const callback = `${this.configService.config.api_base_url}/${this.configService.config.oauth_callback_url}`;
+        const callback = `${this.configService.config.api_base_url}/${
+            this.configService.config.oauth_callback_url
+        }`;
         return `${base}/login?response_type=code&client_id=${clientId}&redirect_uri=${callback}`;
     }
 
@@ -33,7 +36,9 @@ export class AuthService {
         }
         const headers: HeaderObject = {
             Authorization: this.cookieService.get(this.COOKIE_KEYS.TOKEN),
-            [this.COOKIE_KEYS.USER]: this.cookieService.get(this.COOKIE_KEYS.USER),
+            [this.COOKIE_KEYS.USER]: this.cookieService.get(
+                this.COOKIE_KEYS.USER
+            )
         };
         return headers;
     }
@@ -48,7 +53,9 @@ export class AuthService {
 
     isAuthenticated(): boolean {
         const jwt = this.cookieService.get(this.COOKIE_KEYS.TOKEN);
-        if (!jwt) { return false; }
+        if (!jwt) {
+            return false;
+        }
         const jwtData = this.decodeJwt(jwt);
         const expired = jwtData.exp > new Date().getTime();
         // do stuff!!
